prep_local_jupyter.sh: added option to git clone with ssh or https (doesn't try to rsync fastqs in 'https' mode)

+++ reproducibility/prep_local_jupyter.sh
@@ -1,11 +1,18 @@
 set -u
-TARGET_DIR=${1:-hts2018}
+TARGET_DIR=${1:-"hts2018"}
+GIT_METHOD=${2:-"https"}
+
+if [ $# -lt 1 ]; then
+  echo "Usage: $0 TARGET_DIR [GIT_CLONE_METHOD]"
+  echo "       TARGET_DIR: parent directory for downloading data and repos (required)."
+  echo "       GIT_CLONE_METHOD: Use 'ssh' or 'https' for cloning git repos (default: 'https')."
+  exit
+fi
 
 WORK_DIR=$TARGET_DIR/jovyan/work
 DATA_DIR=$TARGET_DIR/data/hts2018_pilot
 JUPYTER_DIR=$TARGET_DIR/jupyter-HTS-2018
 NOTEBOOK_DIR=$WORK_DIR/HTS2018-notebooks
-
 
 rm -rf $NOTEBOOK_DIR $JUPYTER_DIR
 mkdir -p $WORK_DIR $DATA_DIR
@@ -20,9 +27,28 @@ DownloadData() {
     # md5sum -vc Granek_4837_180427A5/Granek_4837_180427A5.checksum | grep -v open
 }
 
+DownloadNotebooks() {
+    if [ $GIT_METHOD == "ssh" ] ; then
+	echo "Cloning HTS2018-notebooks.git with ssh"
+	NOTEBOOK_URL="git@gitlab.oit.duke.edu:HTS2018/HTS2018-notebooks.git"
+    else
+	echo "Cloning HTS2018-notebooks.git with https"
+	NOTEBOOK_URL="https://gitlab.oit.duke.edu/HTS2018/HTS2018-notebooks.git"
+    fi
+    git clone $NOTEBOOK_URL $NOTEBOOK_DIR
+}
+
+
 # Build Jupyter Docker Image
 BuildAndRunImage() {
-    git clone git@gitlab.oit.duke.edu:HTS2018/jupyter-HTS-2018.git $JUPYTER_DIR
+    if [ $GIT_METHOD == "ssh" ] ; then
+	echo "Cloning jupyter-HTS-2018.git with ssh"
+	JUPYTER_GIT="git@gitlab.oit.duke.edu:HTS2018/jupyter-HTS-2018.git"
+    else
+	echo "Cloning jupyter-HTS-2018.git with https"
+	JUPYTER_GIT="https://gitlab.oit.duke.edu/HTS2018/jupyter-HTS-2018.git"
+    fi
+    git clone $JUPYTER_GIT $JUPYTER_DIR
     docker build -t mccahill/jupyter-hts-2018 $JUPYTER_DIR
     echo "docker run --name jupyter-hts-2018 \
       -e USE_HTTPS=yes \
@@ -36,10 +62,16 @@ BuildAndRunImage() {
 }
 
 # # Clone Notebook Repo
-git clone git@gitlab.oit.duke.edu:HTS2018/HTS2018-notebooks.git $NOTEBOOK_DIR
 
-DownloadData
+if [ $GIT_METHOD == "ssh" ] ; then
+    echo "Downloading FASTQs"
+    DownloadData
+else
+    echo "Can't download FASTQs without proper ssh key"
+fi
 
+
+DownloadNotebooks
 BuildAndRunImage
 
 printf "\n\nJUPYTER URL is <https://localhost:9999>\n"
